[BE] [refactor] #3 : PK 값 Long 에서 int 로 변경

+++ backend/src/main/java/com/taffy/backend/poomsae/domain/Mv.java
@@ -18,7 +18,7 @@ public class Mv {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "mv_id")
-    private Long mvId;
+    private Integer mvId;
 
     @Column(name = "mv_url", length = 500)
     private String mvUrl;
